fails cleanly when local files don't exist

+++ socs/db/suprsync.py
@@ -332,17 +332,26 @@ class SupRsyncFileHandler:
             # Creates temp directory with remote dir structure of symlinks for
             # rsync to copy.
             file_map = {}
+            remote_paths = []
             with tempfile.TemporaryDirectory() as tmp_dir:
                 self.log.info("Copying files:")
                 for file in files:
                     self.log.info(f"- {file.local_path}")
                     tmp_path = os.path.join(tmp_dir, file.remote_path)
                     os.makedirs(os.path.dirname(tmp_path), exist_ok=True)
+
+                    if not os.path.exists(file.local_path):
+                        self.log.warn("Cannot find file {path}", path=file.local_path)
+                        file.failed_copy_attempts += 1
+                        continue
+
                     os.symlink(file.local_path, tmp_path)
+
 
                     remote_path = os.path.normpath(
                         os.path.join(self.remote_basedir, file.remote_path)
                     )
+                    remote_paths.append(remote_path)
                     file_map[remote_path] = file
 
                 cmd = ['rsync', '-Lrt']
@@ -355,11 +364,6 @@ class SupRsyncFileHandler:
             for file in files:
                 file.copied = time.time()
 
-            remote_paths = [
-                os.path.join(self.remote_basedir, f.remote_path)
-                for f in files
-            ]
-
             res = self.run_on_remote(['md5sum'] + remote_paths)
             for line in res.stdout.decode().split('\n'):
                 split = line.split()
@@ -370,7 +374,9 @@ class SupRsyncFileHandler:
                     continue
 
                 md5sum, path = line.split()
-                file_map[os.path.normpath(path)].remote_md5sum = md5sum
+                key = os.path.normpath(path)
+                if key in file_map:
+                    file_map[key].remote_md5sum = md5sum
 
             for file in files:
                 if file.remote_md5sum != file.local_md5sum:
